+++ docs/CHANGELOG.rst
@@ -17,6 +17,10 @@ Deprecated
 ----------
 * Drop support for python 3.7
 
+Added
+-----
+- Add support for arm64 architecture on linux desktop (#766)
+
 `v3.1.12`_ (2, Mar. 2024)
 =========================